extracting some zk concerns from send raw transaction call
