+++ cmd/rpcdaemon/commands/send_transaction.go
@@ -5,13 +5,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math/big"
-	"strings"
-
 	"github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/common/hexutility"
 	txPoolProto "github.com/ledgerwatch/erigon-lib/gointerfaces/txpool"
 	"github.com/ledgerwatch/log/v3"
+	"math/big"
 
 	"github.com/ledgerwatch/erigon/core/rawdb"
 	"github.com/ledgerwatch/erigon/core/types"
@@ -19,9 +17,6 @@ import (
 	"github.com/ledgerwatch/erigon/eth/ethconfig"
 	"github.com/ledgerwatch/erigon/params"
 	"github.com/ledgerwatch/erigon/rlp"
-	"github.com/ledgerwatch/erigon/zk/sequencer"
-	"github.com/ledgerwatch/erigon/zk/zkchainconfig"
-	"github.com/ledgerwatch/erigon/zkevm/jsonrpc/client"
 )
 
 // SendRawTransaction implements eth_sendRawTransaction. Creates new message call transaction or a contract creation for previously-signed transactions.
@@ -38,8 +33,7 @@ func (api *APIImpl) SendRawTransaction(ctx context.Context, encodedTx hexutility
 	chainId := cc.ChainID
 
 	// [zkevm] - proxy the request if the chainID is ZK and not a sequencer
-	sequencer := sequencer.IsSequencer()
-	if !sequencer && zkchainconfig.IsZk(chainId.Uint64()) {
+	if api.isZkNonSequencer(chainId) {
 		return api.sendTxZk(api.L2RpcUrl, encodedTx, chainId.Uint64())
 	}
 
@@ -92,24 +86,6 @@ func (api *APIImpl) SendRawTransaction(ctx context.Context, encodedTx hexutility
 	}
 
 	return txn.Hash(), nil
-}
-
-func (api *APIImpl) sendTxZk(rpcUrl string, encodedTx hexutility.Bytes, chainId uint64) (common.Hash, error) {
-
-	res, err := client.JSONRPCCall(rpcUrl, "eth_sendRawTransaction", encodedTx)
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	if res.Error != nil {
-		return common.Hash{}, fmt.Errorf("RPC error response: %s", res.Error.Message)
-	}
-
-	//hash comes in escaped quotes, so we trim them here
-	// \"0x1234\" -> 0x1234
-	hashHex := strings.Trim(string(res.Result), "\"")
-
-	return common.HexToHash(hashHex), nil
 }
 
 // SendTransaction implements eth_sendTransaction. Creates new message call transaction or a contract creation if the data field contains code.
